requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
httpx>=0.27.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
numpy>=1.24.0
